indexing_techniques/multi_vector_indexing/create_vectorstore.py
"""
Load text files, split them into smaller chunks and index those chunks into vectorstore
"""
from langchain.storage import InMemoryByteStore
from langchain_chroma import Chroma
from langchain_community.document_loaders import TextLoader
from langchain_openai import OpenAIEmbeddings
from langchain_text_splitters import RecursiveCharacterTextSplitter
from dotenv import load_dotenv

load_dotenv()

loaders = [
    TextLoader("indexing_techniques/multi_vector_indexing/data/paul_graham_essay.txt"),
    TextLoader("indexing_techniques/multi_vector_indexing/data/state_of_the_union.txt"),
]
docs = []
for loader in loaders:
    docs.extend(loader.load())
text_splitter = RecursiveCharacterTextSplitter(chunk_size=10000)
docs = text_splitter.split_documents(docs)

# The vectorstore to use to index the child chunks
vectorstore = Chroma(
    collection_name="full_documents", embedding_function=OpenAIEmbeddings()
)