indexing_techniques/multi_vector_indexing/pyproject.toml
[tool.poetry]
name = "multi vector indexing"
version = "0.1.0"
description = ""
authors = ["Anantvir Singh"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
chroma = "^0.2.0"
langchain = "^0.2.15"
langchain-openai = "^0.1.23"
langchain-community = "^0.2.15"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
